+++ commands/wizard.js
@@ -5,7 +5,7 @@ module.exports = {
         .setName('wizard')
         .setDescription("Determines if you are the wizard"),
     execute(message){
-        if(message.author.id === '879156695018897409'){
+        if(message.member.roles.cache.has('891540072661397544')){
             message.react('<:wlpog:891123046453227562>');
             message.channel.send(`<:wlpog:891123046453227562> WIZARD <:wrpog:890046143696871474>`);
         }else{