+++ docs/changelog.md
@@ -1,9 +1,15 @@
 # Change Log
 
 ## Version 0.1.3 | 10/8/2021
-- Added `$wizard` command
+- Updated `$wizard` command
+- Updated `$duel` command
+- Updated `$profile` command
+- Updated `$lb` command
+- Updated `$stonks` command
+- Updated `$fruit` command
 
 ## Version 0.1.2 | 9/26/2021
+- Added `$wizard` command
 - Added `$duel` command
 - Added `$profile` command
 - Added `$lb` command